Add tests for Worker.
